Renamed packages to comply with java conventions
