--- src/main/java/POMDPProblem/POMDPDummyProblems/POMDPDummyProblemI.java
@@ -1,7 +1,0 @@
-package POMDPProblem.POMDPDummyProblems;
-
-import POMDPProblem.POMDPProblem;
-
-public interface POMDPDummyProblemI {
-    POMDPProblem load();
-}
