+++ src/main/java/ahsvi/Config.java
@@ -1,4 +1,4 @@
-package AHSVI;
+package ahsvi;
 
 public class Config {
     public static final double ZERO = 1e-6;
