+++ src/main/java/pomdpproblem/pomdpdummyproblems/POMDP1S.java
@@ -1,21 +1,20 @@
-package POMDPProblem.POMDPDummyProblems;
+package pomdpproblem.pomdpdummyproblems;
 
-import POMDPProblem.POMDPProblem;
+import pomdpproblem.POMDPProblem;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 
-public class POMDP2S implements POMDPDummyProblemI {
+public class POMDP1S implements POMDPDummyProblemI {
     @Override
     public POMDPProblem load() {
         // preamble
         double discount = 0.75;
 
+
         ArrayList<String> stateNames = new ArrayList<>();
         HashMap<String, Integer> stateNameToIndex = new HashMap<>();
-
         stateNames.add("s");
-        stateNames.add("v");
         for (int i = 0; i < stateNames.size(); ++i) {
             stateNameToIndex.put(stateNames.get(i), i);
         }
@@ -38,29 +37,23 @@ public class POMDP2S implements POMDPDummyProblemI {
             observationNameToIndex.put(observationNames.get(i), i);
         }
 
-
         // pomdp body
         double[][][] actionProbabilities =
                 new double[actionNames.size()][stateNames.size()][stateNames.size()];
 
-        actionProbabilities[0][0][0] = 0.0;
-        actionProbabilities[0][0][1] = 1.0;
-        actionProbabilities[0][1][0] = 1.0;
-        actionProbabilities[0][1][1] = 0.0;
+        actionProbabilities[0][0][0] = 1.0;
 
 
         double[][][] observationProbabilities =
                 new double[actionNames.size()][stateNames.size()][observationNames.size()];
 
         observationProbabilities[0][0][0] = 1.0;
-        observationProbabilities[0][1][0] = 1.0;
 
 
         double[][][][] rewards =
                 new double[actionNames.size()][stateNames.size()][stateNames.size()][observationNames.size()];
 
-        rewards[0][0][1][0] = 1.0;
-        rewards[0][1][0][0] = 0.0;
+        rewards[0][0][0][0] = 1.0;
 
 
         double[] initBelief = new double[stateNames.size()];
