+++ src/main/java/pomdpproblem/pomdpdummyproblems/POMDPDummyProblemI.java
@@ -1,0 +1,7 @@
+package pomdpproblem.pomdpdummyproblems;
+
+import pomdpproblem.POMDPProblem;
+
+public interface POMDPDummyProblemI {
+    POMDPProblem load();
+}
